Add uint64 cast before shift

+++ gtkWorkspace/sre_can.c
@@ -99,14 +99,14 @@ uint8_t read_and_match()
 			// buffered_data |= (uint64_t)frame.data[5] >> 16;
 			// buffered_data |= (uint64_t)frame.data[6] >> 8;
 			// buffered_data |= (uint64_t)frame.data[7];
-			buffered_data = frame.data[0];
-			buffered_data |= frame.data[1] << 8;
-			buffered_data |= frame.data[2] << 16;
-			buffered_data |= frame.data[3] << 24;
-			buffered_data |= frame.data[4] << 32;
-			buffered_data |= frame.data[5] << 40;
-			buffered_data |= frame.data[6] << 48;
-			buffered_data |= frame.data[7] << 56;
+			buffered_data = (uint64_t)frame.data[0];
+			buffered_data |= (uint64_t)frame.data[1] << 8;
+			buffered_data |= (uint64_t)frame.data[2] << 16;
+			buffered_data |= (uint64_t)frame.data[3] << 24;
+			buffered_data |= (uint64_t)frame.data[4] << 32;
+			buffered_data |= (uint64_t)frame.data[5] << 40;
+			buffered_data |= (uint64_t)frame.data[6] << 48;
+			buffered_data |= (uint64_t)frame.data[7] << 56;
 
 			printf("Frame ID: %d,BufferData %ld\n",frame.can_id, buffered_data);
